Unify server spawns
With #2 the low-level `spawnHttpServer()` operation was introduced to
make spawning with any http interface work. It was based on the
original spawnSimulationServer(), but this makes the original
simulation server use it.

With that in place, and with the actual server managing the servers,
we don't need any other type of http server than express, so we can
simplify the types.