Add a person simulator
We need to do more than just bind services. We have to be able to
declare scenarios that arrange the state of the simulation into a
specific order.

This addse the concept of a "scenario" to the simulation and the
given() GraphQL API that can be used to add it to the current
simulation.

As the first implementer of this scenario API is the "person"
simulator which currently exports a single scenario called "person"
which generates the person and inserts it into the store.

The person generation logic is deliberately hokey and we can upgrade
it to use faker / fast-check, or some combination later on.