drive entire simulation as an effect mapped off simulation state
Since we're about the business of adding the atom to track simulation
with https://github.com/thefrontside/simulacrum/pull/6 we have an
opportunity to completely re-think how this will work and organize
accordingly.

This implements the idea that the entire simulation is tracked in the
atom from the moment it is created until the moment it is
destroyed. Instead of doing everything in the graphql context, the
context just inserts the initial simulation record into the atom and
the `map()` effect which maps each record to an operation takes it
away from there.