Remove builder syntax
The builder syntax is getting in the way of rapid innovation. Rather
than iterating on the simulation server and quickly adding the
scenarios feature, I find that I'm fighting the builder syntax that I
was advocating for in the beginning 🤦🏼 The idea was that there would
be a nice chaining builder syntax that would help guide you on the
right path, but right now it feels like a millstone and this is early
days yet.

This rolls back pretty much anything fancy. The job of a builder
object is to emit a data structure, and so this just replaces that to
allow simulators to provide the data structure directly.

Before

```

export const simulator = (behaviors) =>
  behaviors.
    http(app =>
      app.get('/', echo))
    .https(app => app.get('/', echo), 'secure-echo')
```

After

```
import { createHttpApp } from 'simulacrum';

export const simulator = () => ({
  services: {
    echo: {
      protocol: 'http',
      app: createHttpApp().get('/', echo)
    },
    'secure-echo': {
      protocol: 'https',
      app: createHttpApp().get('/', echo)
    }
  }
})
```

This is only slightly more verbose, but it is a hell of a lot easier
to implement, and if we want some fancy way to build up those data
structures, we can always add them later. But as-is, they are only
getting in the way.