Add option to make predictable simulation sequence
In order to make testing and development easier, we wanted the ability
to pass in a uuid to make sure that the created simulation id was
stable so that queries we had sitting around would continue to work
and sequences of snippets used for testing would not be one-use and
obsolete.

However, we can accomplish the same goal by just using a predictable
sequence of UUIDs. That way, when we run a server, and start a
simulation, it always gets the same UUID.

Currently, faker does not support independent instances that each have
their own pseudo random number generators so this change in its
current form depends on an open PR I have on the faker.js repo making
multiple faker instances safe.

We'll need this very soon since each simulation will have its own
random seed and its own faker instance.