✨Introduce standalone ui application
We plan to use [graphql-ws][1] as the primary transport for the
simulacrum server. There are a lot of good reasons for this, but the
most important are:

1. we don't have to implement a protocol ourselves
2. The protocol is cross platform and there are clients for many
different languages
3. As the ecosystem evolves, more tools will form around this protocol.

However, after much digging, it became clear that the graphiql that
ships with `express-graphql` is stale and does not support the more
recent version of the protocol. There are several issues open on
github to upgrade the version of graphiql that ships with, none of
which have received attention in a long time. In fact, it appears that
the express-graphql project, while not quite abandoned, is not very
actively maintained.

In order to use a modern graphiql, so that we can implement our own
fetcher, we have to roll our own app that uses the graphiql react
component. Rather than vendoring graphiql, this bundles a complete app
with parcel that contains graphiql, and serves that app from the
server.

While this is currently in the server package for now, this change is
made with the understanding that it will be extracted it into it's own
package, `@simulacrum/ui` when we have npm publishing set
up. Hopefully in time, the UI can grow to be more than just a GraphiQL
instance, but can grow to be a more featureful UI of which one tab is
just the graphiql instance.

To that end, it is possible to run the UI, _not_ served by the
server. You can currently pass the server url as a parameter to the UI
if you want to hack on the UI by itself with the `?server=` query
param. In that case you would start your server, and then start the UI
separately with a call to:

````
> parcel server ui/index.html --target ui
```

And then navigating to
`http://localhost:1234?server=http://localhost:3000`

[1]: https://github.com/enisdenjo/graphql-ws