Break apart graphql context
Integrating websockets requires some pretty big changes to the way we
execute graphql queries. Specifically, the Effection scope will be
different when we do queries over http than when we do queries over
websockets. The scope of the http requests will be shared for all
requests, but for the subscription we'll want a scope that is
per-websocket. Because of this, we won't be able to use a one size
fits all graphql context.

This breaks apart the context, which was more of a "resolver holder"
into a bunch of free standing functions that take the bare minimum of
things which they need to run: an effection context, an atom, and an
id sequence.

The schema can now just include these free standing functions by
passing through the context. It is up to the http and wss layer to
choose the appropriate effection scope.