Check that index.html was found, not just the directory.
We have a sanity check in place, but with very little effort, it could
be enhanced to be more sane.

Instead of checking that the `dist` directory exists, this checks that
the actual `dist/index.html` itself exists. It's possible to have
cleaned out the build directory, but left it in place.