Fix global clean
`npm run clean` from the project root will not actually clean
everything up. Also, running a build is awkward since we have to `cd`
into each package and invoke the build script.

Luckily, with npm 7.7.0, we have the capability to fan out a command
over all the workspaces by passing the `-ws` flag. This updates the
`clean`, `build` and `test` commands to use this functionality so that
each project can decide for itself what it means to do these things.