Pass data generation API to scenarios
Right now, we don't provide any support to our scenarios for
generating data. That means that each individual scenario that need to
generate data is left completely to its own devices. But given that
this is going to be very common, this is something that the simulation
server should provide out of the box.

This creates a single instance of faker per simulation, and passes it
into each scenario so that the scenario can use it to generate
high-fidelity simulation data.

As a bonus, I was able to actually get rudimentary typing on the
`Faker` interface. However, the `@types/faker` module is quite behind
the actual faker and so there are a lot of methods on the faker
instance that aren't represented in the typings. Given that caveat, it
seems to me better than just having it typed as `any`.