🐛Point to ui app dir, not ui index.html
With https://github.com/thefrontside/simulacrum/pull/25 we put in a
check to see that `dist/index.html` was present, not just the `dist/`
directory. I screwed that up by also returning that as the value of
`appDir()`. So while it did exist, what the express server needs is
the application root directory.

If we can get our bigtest harness up and running for this, it would
have caught this.