✨@simulacrum/client
One of the super-powers of the simulacrum simulation server is that it
has a GraphQL API that can be called from multiple different
platforms, such as JavaScript, .Net, Go, Rust, etc... However, it is
more convenient not to call through to the low-level graphql API, but
instead to use a native language API.

This introduces the first such language API for JavaScript, and it is
built to be an all-purpose JS client for Simulacrum. As such, it does
not expose any effection API, but instead uses `Promise`, and
`AsyncIterable` externally, even though it uses effection underneath
to implement resource disposal and request cancellation.