♻Convert all tests to use the JavaScript Client
While it's cool that the control API is graphql. It is soooooo much
more convenient to use the Javascript client instead. Not only is it
less code, it typechecks much better.

This converts all of the tests to use the official client. This means
that we are able to convert the `createTestServer` helper to yield an
instance of the `Client` itself to make tests very consise. It now
creates a client / server pair.