Check in package-lock.json
We want our builds to be repeatable and not fail for random reasons
not related to our code changes. If a dependency changes that could be
installed by our package dependency ranges that causes a failure, we
don't want that to fail a build randomly.

At the same time, we don't want to spending the bulk of our time in
our build calculating and then downloading dependencies. Instead we
should be able to cache a known set of dependencies and spend most of
our resources actually running the build and tests.

This checks in the package-lock.json so that we get the same
dependencies each time and our build can be sped up.

There is a drawback in that dependencies could introduce bugs that our
test suite could detect and we would never know of them because we'd
never run our test suite against them. However, in those cases the
answer is to make a "edge" build that doesn't block merges but servers
as a canary in the coalmine for impending changes that make cause
things to break.