Relax internal dependencies so that they still match.
There is a special case in semver where `0` level caret dependencies
are equivalent to pins. This is because the common use-cases is that
while developing in the 0 range, even minor changes are
breaking. This means that a caret range like `^0.0.1` is not
satisfied by `v0.1.0`, even though `^1.0.1` _would_ be satisfied by
`1.1.0`.

As a result, when our release PR advances the version of the client to,
it becomes immediately incompatible with the server dependency of the
server, and so npm wants to install the old version from NPM rather
than symlinking to the package in the repository. In a way, this isn't
wrong behavior, but it is suprising that npm doesn't even issue a
warning when it does this.

This pins the versions to internal dependencies