✨destroy simulations
Simulations are big stateful things, and especially when they are part
of the testcases we need to be able to create and destroy them
instantaneously. Whenever a simulation is destroyed, any effects
associated with it should be torn down.

This removes the slice associated with the simulation, and since all
the effects are derived off of that slice, they are immediately shut
down.

Note: because of some weirdness with our atom implementation, the
removal is somewhat weird in that it results in the effect hook
getting fired with the state being `undefined`, and also with an
undefined value. While this should technically never happen, it
does. and so we have to work around it until a more permanent solution
can be found.