✨Type-check tests
Before the tests were not included in the tsconfig.json because they
are not something that we want to distribute. However, we do want them
to show errors in editors and also be able to typecheck them
statically.

This adds the test directory to the base config, and then creates a
special "dist" config for compiling the distributable sources.