Make a proper release
Right now, we're just bundling up the source of Simulacrum and making
it downloadable on NPM, but as an executable program, it is a
full-featured system written in TypeScript, JavaScript, HTML and
CSS. As such, we need to make sure and build all of our sources into
their target executables.

This adds a prepack script to each package that builds its local
sources and makes it ready to distribute to npmjs.com. Among other
things we:

1. build typescript into javascript
2. add at least a minimal README to every package
3. Fix the author field to not be just cowboyd personally, but all of
FS engineering
4. Only distribute built files, not sources