Only one top-level simulator in a simulation
We want to add the concept of simulation parameters. For example, if I
have multiple versions of an API, then I might want to have a "stage"
parameter that I pass to `createSimulation()`.

```js
createSimulation(["gateway"], {
  stage: "production"
});
```

This begs the question, which simulator would receive this if I was
creating the simulation with multiple simulators?

```js
createSimulation(["auth0", "gateway"], {
  stage: "production"
})
```

It seems very unclear when it shouldn't be. There's a need to pass the
properties to both and then let them just do what they want. but that
seems very odd indeed.

We can avoid this situation entirely by saying that you must have a
single simulator sitting at the root of your tree, similar to the way
that you have to have a single React component that you render in the
DOM. In the same way that the React component can be comprised of
several different child compoents, so we could also have a root
simulator that was comprised of multiple different simulator
"children".

This takes a first step in that direction by restricting
createSimulation to a single simulator. That way, if you want to
compose `auth0` and `backend`, you would need to create a root
simulator that would compose them explicitly rather than implicitly:

```
-> backend
   -> auth0
   -> gateway
```

and you would instantiate it like so:

```js
createSimulatation("backend", {
  stage: "production"
})

Where the backend simulation would start _both_ `auth0` _and_
`gateway` and explicitly delegate the `stage` parameter to gateway.

The APIs for composing simulators are yet to be settled, but the core
concept, that you have a  "component-like" tree that rolls up
into a single simulator is still the basic idea.

This makes the implementation quite a bit simpler, where instead of
having to merge and combine scenarios and services from the different
simulators, the composition happens at the simulator level.