Debounce watch script
When I was reviewing the watch script it occured to me that the
debounce logic wasn't right. The delay needs to go at the beginning of
the process. That way, the first thing it does is `sleep()` for 50ms
so if any subsequent watch events come in, the current will be
sleeping and get cancelled before it even does anything.