Remove covector from development dependencies
When trying to upgrade to effection 2.0 TypeScript was incorrectly
loading the type definitions for 1.0 that were brought in by
covector. This is a bug in our TypeScript config, but it's one that we
don't have to deal with right now since we don't strictly need
covector in our development dependencies, since we use the github
action for it in CI.

This just removes covector from the dev dependencies, so that we can
freely upgrade effection. This kicks the problem down the road of what
to do with our tsconfig that it is resolving modules incorrectly, but
would be nice to not find TS for awhile.