Get rid of `webSocketImpl` argument in `createClient`
It's annoying and confusing that on node you have to explicitly pass
in your WebSocket constructor to the `createClient` method. We want to
have the same signature of `createClient(url: string)` on all
platforms and do a good job of detecting and choosing the websocket
library that is appropriate.

In this case, we just always explicitly pass the websocket constructor
to the underlying `graphql-ws` instance using `isomorphic-ws` since it
is created for exactly this situation.