add basic logging middleware (#135)
## Motivation

We need more visibility as to what is going on in simulators like auth0 which are annoying to debug.

## Approach

There is a basic middleware component in `@simulacrum/server` that simulators can use to log HTTP requests.

It can take 1 or more logging function.  I've add one logging function to auth0 that writes to the affection inspector:

```ts
app: createHttpApp()
      .use(createLoggingMiddleware(function * (message) {
        yield label({ name: 'auth0-logging-middleware', log: message });
      }))
```

It should be easy to add other loggers like `console.log` for quicker debugging.