Allow LDAP server to be standalone a resource
In the article about running ingestion tests for backstage, we need to
run a simulated LDAP server outside the context of simulacrum. The
primary reason for this is that we do not need yet to introduce the
concept of simulacrum as its own thing. It just complicates the
world.

This just splits the LDAP simulator into two layers, the server
resource, created by `createLDAPServer` and the service resource that
adapts it into the Simulacrum context `createLDAPService()`. The user
data is represented as an `Iterable` which is the lowest level JS
format for specifying a list of values. It could be an arry, or it
could be something more.