Implement root DSE
Every LDAP Server must have a [Root DSE][1] to provide metadata about
itself. This allows clients to distinguish between particular
vendors. Among others, Backstage uses the root dse to determine if it
is talking to Microsoft Active Directory and customize its behavior
accordingly.

This adds the root DSE search which is just an "empty" entry with the
zero leng string as its DN.