Simplify the LDAP schema
There are some difficulties integrating the LDAP simulator in with the
Backstage processors and providers. The `UserData` that feeds the LDAP
server is required to have an `id` attribute and an `email` attribute,
both of which are not part of any common ldap schema. Furthermore, it
does not send common user attributes as part of a search result. So,
for example, Backstage (as well as most common LDAP clients), when
connecting to an OpenLDAP server, will be looking for both the
`entryDN` virtual attribute and the `entryUUID` attribute. Backstage,
for example, uses `entryUUID` as the value of the all-important
`metadata.name` Entity field.

This aligns the `user data` with the minimum number of
attributes. Since `cn` is used to authenticate, we use `cn` in the
user data api (also, the LDAP convention for the email field is `mail`
anyway, and not `email`). This also makes send the the `dn` back to
the client more straightforward since the `dn` is comprised of the
`cn`,`baseDN`.

Note that this is a breaking change, and will require user data to be
mapped differently for currently running these simulators.