Prep API and README for testing article (#200)
* Prep API and README for testing article

There is an article in the works that references the LDAP simulator as
a solution to testing a backstage server. Unfortunately, we don't have
a README, and so the [NPM package
homepage](https://www.npmjs.com/package/@simulacrum/ldap-simulator/v/0.4.0)
looks quite bad.

This adds a basic readme and shows how to start an LDAP server from
both a vanilla JavaScript context as well from an Effection
context. There is a third way, which is via a simulacrum server, but
this does not add that since it's a bit more fiddly, likely to change,
and not relevant for the content of the blog post.

It also adds the actual vanilla javascript integration since it didn't
exist before (it just wraps the resource in a run);

* Code formatting was broken

Co-authored-by: Taras Mankovski <taras@frontside.com>