✨Add ability to extende the Github API Router with custom middleware
Our implementation of the GitHub API is incomplete which is not
suprising given that it is huge. However this poses a problem for
someone using the API simulator. What if the API is missing something
that they need? Their options are to submit a PR to this repo, or to
just start the things they need on another port. This is doable, but
it means that they have to manage two servers, and also that the
simulation does not appear to be hosted at the same origin. That's an
unecessary burden. We need a way to extend a simulator with new
endpoints (and possibly new middlewares) without requiring an upstream
change to this repository.

This takes the simplest approach that could possibly work. It adds an
"extension" property to the `ServerOptions` interface. It's just a
function that, if present, will be invoked with the express router so
that it can do whatever it wants to it.